🐛 Fix: Uses a tab instead of spaces for inquirer
In the original design, we used two spaces to pad the emoji. Unfortunately, many emoji are now multi-character width and can cause alignment issues in the inquirer prompt list. By changes from spaces to a tab character, we can ensure a clean alignment in the inquirer list without affecting the commit itself.